+++ tests/test_flake8_async.py
@@ -476,7 +476,7 @@ error_codes_ignored_when_checking_transformed_sync_code = {
     "ASYNC116",
     "ASYNC117",
     "ASYNC118",
-    "ASYNC120",
+    "ASYNC300",
     "ASYNC912",
 }